feat(bot): move owner premium commands to admin
